refactor: Line에게 필요 없는 파라미터 삭제 후, 제약조건을 Lines로 이전
