+++ src/main/java/nextstep/ladder/domain/Lines.java
@@ -5,13 +5,26 @@ import java.util.List;
 
 public class Lines {
 
+    public static final int MINIMUM_PARTICIPANTS_COUNT = 2;
     private final List<Line> lines;
 
     private Lines(Floor floor, int countOfUsers, LineStrategy strategy) {
+        validate(countOfUsers, strategy);
+
         this.lines = new ArrayList<>();
 
         for (int i = 0; i < floor.getFloor(); i++) {
             lines.add(strategy.execute(countOfUsers));
+        }
+    }
+
+    private void validate(int countOfUsers, LineStrategy strategy) {
+        if (countOfUsers < MINIMUM_PARTICIPANTS_COUNT) {
+            throw new IllegalArgumentException(String.format("사용자 수는 %d명 이상이여야 합니다.", Lines.MINIMUM_PARTICIPANTS_COUNT));
+        }
+
+        if (countOfUsers != (strategy.execute(countOfUsers).size() + 1)) {
+            throw new IllegalArgumentException("사용자 수와 사다리 전략 결과의 검증에 실패했습니다.");
         }
     }
